further refinements and optimizations
Mainly a load of housekeeping - removed tabs for spaces in the indentation, re-ordered some defines, added a load of comments and put all the functions after the loop function, to make it more like other arduino code I've seen. Must be best practice or something.
More significant changes were the removal of the lightNum element of the light structure, as it wasn't being used, changed requiredState variable to use OFF and ON rather than LOW and HIGH (made more sense when reading the code).
Finally removed some unnecessary local variables in a couple of the structures (level in lightLED and currentTime in setup_wait_period - currentTime is already a global)