Time trials
Added code to time and display the average execution times of selected algorithms. Also added several different algorithms for each bit under test to see which was fastest and smallest. Results will influence the main code going forward but the timing code will not carry on in to the main branch.