Block further changes whilst changing
The read of the LDR input is now not done if the system is already in the middle of changing the state of the lights. This is to prevent it getting in to a mess if the LDR input is rapidly changing (though in theory it shouldn't).
Also changed the PWM_VARIANCE from 512 to 384 and the min flicker time from 75 to 100, just to see how it looks.